// src/State/clinkRequests/types.ts
import { LiveDebitRequest, LiveManageRequest } from "@/Api/pub/autogenerated/ts/types";
import { NprofileView } from "@/State/scoped/backups/sources/selectors";

export type PendingClinkDebitRequest = {
	kind: "debit";
	identityId: string;
	sourceId: string;
	request: LiveDebitRequest;
	enqueuedAtMs: number;

};

export type PendingClinkManageRequest = {
	kind: "manage";
	identityId: string;
	sourceId: string;
	request: LiveManageRequest;
	enqueuedAtMs: number;
}

export type PendingClinkRequest = PendingClinkDebitRequest | PendingClinkManageRequest;

export type PendingClinkDebitSession = {
	kind: "debit";
	request: LiveDebitRequest;
	source: NprofileView;
}

export type PendingClinkManageSession = {
	kind: "manage";
	request: LiveManageRequest;
	source: NprofileView;
}

export type PendingClinkSession = PendingClinkDebitSession | PendingClinkManageSession;

export interface ClinkRequestsState {
	pendingClinkRequests: PendingClinkRequest[];
	pendingClinkRequestSession: PendingClinkSession | null;
}